made final adjustments

+++ routes/admin.js
@@ -4,10 +4,10 @@ var conn = require('../lib/db')
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
-    if (req.session.loggedin == true){
-            
-            res.render('admin/index', { title: 'DashBoard', my_session : req.session });
+    if (req.session.loggedin == true && req.session.role === 'yes'){    
+         res.render('admin/index', { title: 'DashBoard', my_session : req.session });
     }else {
+        req.flash('error', ('error'))
         res.redirect('/login')
     }
 });
@@ -15,13 +15,14 @@ router.get('/', function(req, res, next) {
 router.get('/view_employee', (req, res) => {
     
     let sql = 'SELECT * FROM pet_store.admin'
-    if (req.session.loggedin == true) {
+    if (req.session.loggedin == true && req.session.role === 'yes') {
         conn.query(sql, (err, results) => {
             if (err) {
                 res.render('admin/admin_info/admin_view',
                 {
                     title: 'Employees List',
-                    employees: ''
+                    employees: '',
+                    my_session : req.session
                 })
             }else {
                 res.render('admin/admin_info/admin_view',
@@ -39,8 +40,11 @@ router.get('/view_employee', (req, res) => {
 
 // Get Add Admin
 router.get('/add_employee', (req, res) => {
-    if (req.session.loggedin == true) {
-        res.render('admin/admin_info/add_admin', {title: 'Add Employee'})
+    if (req.session.loggedin == true && req.session.role === 'yes') {
+        res.render('admin/admin_info/add_admin', {
+            title: 'Add Employee',
+            my_session : req.session
+        });
     }else {
         res.redirect('/login')
     }
@@ -55,7 +59,7 @@ router.post('/add_employ', (req, res) => {
         email : req.body.email,
         password : req.body.password 
     };
-    if (req.session.loggedin == true) {
+    if (req.session.loggedin == true && req.session.role === 'yes') {
         conn.query(sql, data, (err, results) => {
             if(err) {
                 req.flash('err', 'Error, Try Again');
@@ -72,7 +76,7 @@ router.post('/add_employ', (req, res) => {
 
 // Delete Employee
 router.get('/delete/:id', (req, res) => {
-    if (req.session.loggedin == true) {
+    if (req.session.loggedin == true && req.session.role === 'yes') {
         conn.query('DELETE FROM pet_store.admin WHERE id=' + req.params.id, (err, results) => {
             if (err) {
                 req.flash('error', 'Could Not Delete')
@@ -89,7 +93,7 @@ router.get('/delete/:id', (req, res) => {
 
 // Get Admin Edit
 router.get('/edit/:id', function(req, res, next) {
-    if (req.session.loggedin == true) {
+    if (req.session.loggedin == true && req.session.role === 'yes') {
         conn.query('SELECT * FROM pet_store.admin WHERE Id='+ req.params.id, function(err, results)     {
             if(err){
                 req.flash('error', err); 
@@ -103,6 +107,7 @@ router.get('/edit/:id', function(req, res, next) {
                 {
                     title: "Employee Edit",
                     employees: results[0],
+                    my_session : req.session
                 });
             }               
         });
@@ -120,7 +125,7 @@ router.post('/update', (req, res) => {
         email : req.body.email,
         password : req.body.password 
     };
-    if (req.session.loggedin == true) {
+    if (req.session.loggedin == true && req.session.role === 'yes') {
         conn.query('UPDATE pet_store.admin SET ? WHERE id =' + req.body.id, data, (err, results) => {
             if (err) {
                 req.flash('err', 'Could Not Update');
@@ -128,6 +133,33 @@ router.post('/update', (req, res) => {
             }else {
                 req.flash('success', 'Employee Updated');
                 res.redirect('/admin/view_employee');
+            }
+        });
+    }else {
+        res.redirect('/login')
+    }
+});
+
+// Get Sales
+router.get('/view_sales', (req, res) => {
+    
+    let sql = 'SELECT cs.*, pt.product_name, cust.fName, cust.lName FROM pet_store.product pt, pet_store.customer cust, pet_store.customer_sale cs WHERE cs.customer_id = cust.id AND cs.product_id = pt.id ORDER BY id'
+    if (req.session.loggedin == true && req.session.role === 'yes') {
+        conn.query(sql, (err, results) => {
+            if (err) {
+                res.render('admin/sales',
+                {
+                    title: 'Sales List',
+                    sales: '',
+                    my_session : req.session
+                })
+            }else {
+                res.render('admin/sales',
+                {
+                    title: 'sales List',
+                    sales: results,
+                    my_session : req.session
+                })
             }
         });
     }else {
